feat: 1.修复文档跳转路径 2.新增绘制提示popup 3.拆分utils模块
