+++ docs/基础绘制/pointDrawer/helper.md
@@ -1,0 +1,10 @@
+---
+title: 提示文案
+order: 8
+group: 
+  path: /PointDrawer
+  title: PointDrawer 绘制点
+  order: 1
+---
+
+<code src="./helper.tsx" compact="true" defaultShowCode="true"></code>
