+++ src/constants/helper.ts
@@ -1,0 +1,7 @@
+import { IPointHelper } from '../typings';
+
+export const DEFAULT_POINT_HELPER: IPointHelper = {
+  draw: '单击绘制点',
+  pointHover: '可拖拽点位置',
+  pointDrag: '松开鼠标完成拖拽',
+};
